+++ src/web/css/fwjs-davenport.pcss
@@ -258,3 +258,17 @@ nav ul li button.inactive {
     display: none;
 }
 
+#actions div {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 1rem;
+}
+
+#actions div button {
+    background-color: yellow;
+    border-radius: 1rem;
+    font-size: 1.5rem;
+    padding: .5rem 1rem;
+    border: 1px inset black;
+}
