+++ src/web/index.html
@@ -14,7 +14,6 @@
                 <h1>Davenport Card Game</h1>
                 <aside>Brought to you by <a href="https://funwithjavascript.com" rel="noreferrer">Fun With JavaScript</a></aside>
             </header>
-            <button id="go">go</button>
             <!--<nav>
                 <ul id="actions">
                     <li><button type="button" id="start-round">Start Round</button></li>
@@ -25,6 +24,11 @@
             </nav>-->
             <section id="playing-field">
                 <section id="opponent-list">
+                </section>
+                <section id="actions">
+                    <div class="">
+                        <button class="new-game">New Game</button>
+                    </div>
                 </section>
             </section>
             <!--<textarea id="results"></textarea>-->
