+++ src/web/js/davenportConfig.js
@@ -1,0 +1,12 @@
+import MovingCard from "./MovingCard.js";
+
+export const playersConfig = [
+	{ id: 'You', human: true },
+	{ id: 'Waldo', avatar: 'img/waldo.webp' },
+	{ id: 'Clarice', avatar: 'img/clarice.webp'  },
+	{ id: 'Bert', avatar: 'img/bert.webp' }
+];
+
+export const animationTimingConfig = {
+		timing: { duration: 250, iterations: 1 }
+}
